:recycle::zap: Cache api servers as well
