+++ tests/conftest.py
@@ -5,8 +5,8 @@ import random
 import pytest
 import torch
 from llm_cache import SortKey, sort_tests_for_llm_caching
-from spyre_util import (RemoteOpenAIServer, clear_cached_llm,
-                        skip_unsupported_tp_size)
+from spyre_util import (clear_cached_llm, clear_cached_runtime_server,
+                        get_cached_api_server, skip_unsupported_tp_size)
 from vllm.connections import global_http_connection
 from vllm.distributed import cleanup_dist_env_and_memory
 
@@ -176,8 +176,10 @@ def remote_openai_server(request):
 
     if 'tp_size' in params:
         tp_size = params['tp_size']
-        skip_unsupported_tp_size(int(tp_size), backend)
-        server_args.extend(["--tensor-parallel-size", str(tp_size)])
+        if int(tp_size) > 1:
+            # Don't set tp size explicitly if it's 1
+            skip_unsupported_tp_size(int(tp_size), backend)
+            server_args.extend(["--tensor-parallel-size", str(tp_size)])
 
     if "cb" in params and params["cb"] == 1:
         max_model_len = params["max_model_len"]
@@ -208,11 +210,25 @@ def remote_openai_server(request):
         }
 
     try:
-        with RemoteOpenAIServer(model, server_args,
-                                env_dict=env_dict) as server:
-            yield server
+        server = get_cached_api_server(model,
+                                       server_args=server_args,
+                                       server_env=env_dict)
+        yield server
+        # with RemoteOpenAIServer(model, server_args,
+        #                         env_dict=env_dict) as server:
+        #     yield server
     except Exception as e:
         pytest.fail(f"Failed to setup server: {e}")
+
+
+@pytest.fixture(scope='session', autouse=True)
+def teardown_fixture():
+    # Session scoped fixture will run once for the entire suite
+    yield
+
+    # Clear out any cached LLMs so no subprocesses get orphaned
+    clear_cached_llm()
+    clear_cached_runtime_server()
 
 
 @pytest.fixture
