+++ tests/llm_cache.py
@@ -1,10 +1,18 @@
 """Contains utilities for caching models (instantiated as vLLM endpoints) 
 across test cases, to speed up test runtime."""
 
+import os
+import subprocess
+import sys
+import time
 from typing import NamedTuple, Optional
 
+import openai
 import pytest
+import requests
 from vllm import LLM
+from vllm.entrypoints.openai.cli_args import make_arg_parser
+from vllm.utils import FlexibleArgumentParser, get_open_port
 
 
 def sort_tests_for_llm_caching(items: list) -> None:
@@ -280,6 +288,182 @@ class LLMCache:
             from spyre_util import force_engine_shutdown
             force_engine_shutdown(self._llm)
             self._llm = None
+            self._runtime_config = None
+
+
+class RemoteOpenAIServer:
+    """Subprocess wrapper that boots a vllm server with `vllm serve` for testing
+    against"""
+
+    DUMMY_API_KEY = "token-abc123"  # vLLM's OpenAI server does not need API key
+
+    def __init__(self,
+                 model: str,
+                 vllm_serve_args: list[str],
+                 *,
+                 env_dict: Optional[dict[str, str]] = None,
+                 seed: Optional[int] = 0,
+                 auto_port: bool = True,
+                 max_wait_seconds: Optional[float] = None) -> None:
+        # NB: This implementation does not ensure that the model is downloaded
+        # before booting the server, it should be used with models already
+        # cached on disk
+
+        if auto_port:
+            if "-p" in vllm_serve_args or "--port" in vllm_serve_args:
+                raise ValueError("You have manually specified the port "
+                                 "when `auto_port=True`.")
+
+            # Don't mutate the input args
+            vllm_serve_args = vllm_serve_args + [
+                "--port", str(get_open_port())
+            ]
+        if seed is not None:
+            if "--seed" in vllm_serve_args:
+                raise ValueError("You have manually specified the seed "
+                                 f"when `seed={seed}`.")
+
+            vllm_serve_args = vllm_serve_args + ["--seed", str(seed)]
+
+        parser = FlexibleArgumentParser(
+            description="vLLM's remote OpenAI server.")
+        parser = make_arg_parser(parser)
+        args = parser.parse_args(["--model", model, *vllm_serve_args])
+        self.host = str(args.host or 'localhost')
+        self.port = int(args.port)
+
+        env = os.environ.copy()
+        if env_dict is not None:
+            env.update(env_dict)
+        self.proc = subprocess.Popen(
+            ["vllm", "serve", model, *vllm_serve_args],
+            env=env,
+            stdout=sys.stdout,
+            stderr=sys.stderr,
+        )
+        max_wait_seconds = max_wait_seconds or 600
+        self._wait_for_server(url=self.url_for("health"),
+                              timeout=max_wait_seconds)
+
+    def __enter__(self):
+        return self
+
+    def __exit__(self, exc_type, exc_value, traceback):
+        self.shutdown()
+
+    def shutdown(self):
+        self.proc.terminate()
+        try:
+            self.proc.wait(8)
+        except subprocess.TimeoutExpired:
+            # force kill if needed
+            self.proc.kill()
+
+    def _wait_for_server(self, *, url: str, timeout: float):
+        # run health check
+        start = time.time()
+        while True:
+            try:
+                if requests.get(url).status_code == 200:
+                    break
+            except Exception:
+                # this exception can only be raised by requests.get,
+                # which means the server is not ready yet.
+                # the stack trace is not useful, so we suppress it
+                # by using `raise from None`.
+                result = self.proc.poll()
+                if result is not None and result != 0:
+                    raise RuntimeError("Server exited unexpectedly.") from None
+
+                time.sleep(0.5)
+                if time.time() - start > timeout:
+                    raise RuntimeError(
+                        "Server failed to start in time.") from None
+
+    @property
+    def url_root(self) -> str:
+        return f"http://{self.host}:{self.port}"
+
+    def url_for(self, *parts: str) -> str:
+        return self.url_root + "/" + "/".join(parts)
+
+    def get_client(self, **kwargs):
+        if "timeout" not in kwargs:
+            kwargs["timeout"] = 600
+        return openai.OpenAI(
+            base_url=self.url_for("v1"),
+            api_key=self.DUMMY_API_KEY,
+            max_retries=0,
+            **kwargs,
+        )
+
+    def get_async_client(self, **kwargs):
+        if "timeout" not in kwargs:
+            kwargs["timeout"] = 600
+        return openai.AsyncOpenAI(base_url=self.url_for("v1"),
+                                  api_key=self.DUMMY_API_KEY,
+                                  max_retries=0,
+                                  **kwargs)
+
+
+class RemoteOpenAIServerCache:
+
+    def __init__(self):
+        self._api_server: RemoteOpenAIServer | None = None
+        self._runtime_config: dict | None = None
+        self._past_runtime_configs: list[dict] = []
+
+        self.hits = 0
+        self.misses = 0
+
+    def __del__(self):
+        if self._api_server:
+            self._api_server.shutdown()
+
+    def get_api_server(self, model: str, server_args: list[str],
+                       server_env: dict) -> RemoteOpenAIServer:
+        """Get or create a new OpenAI server for a given model. and config"""
+        runtime_config = {
+            "model": model,
+            "server_args": tuple(server_args),
+            "server_env": server_env,
+        }
+
+        if self._runtime_config and self._runtime_config == runtime_config:
+            self.hits += 1
+            return self._api_server
+
+        # cache miss
+        self.misses += 1
+
+        print("\n\n\n\n\t\t\tCACHE MISS!\n")
+        print(runtime_config)
+        print()
+        print(self._runtime_config)
+        print("\n\n\n\n")
+
+        assert runtime_config not in self._past_runtime_configs, \
+            f"Runtime config {runtime_config} was previously cached, error " \
+                "in test ordering!"
+
+        self._runtime_config = runtime_config
+        self._past_runtime_configs.append(self._runtime_config)
+
+        # Tear down old server before making a new one
+        if self._api_server:
+            self._api_server.shutdown()
+
+        # Boot up new server
+        self._api_server = RemoteOpenAIServer(model=model,
+                                              vllm_serve_args=server_args,
+                                              env_dict=server_env)
+
+        return self._api_server
+
+    def clear(self) -> None:
+        if self._api_server:
+            self._api_server.shutdown()
+            self._api_server = None
             self._runtime_config = None
